Ratio ml

See merge request jal04/local-search!2

+++ gotzilla.py
@@ -6,8 +6,7 @@ import solvers.single as single
 import solvers.frontier as frontier
 
 
-import pandas as pd 
-import numpy as np 
+import json
 
 import sys
 import os
@@ -16,27 +15,51 @@ import random
 
 def train():
     
-    # Generate CNFs
-    os.system("rm -r benchmark-folder/*")
+
+    def read_tmp_rating(ratio ,dic):
+        ratio_dic = dic.get(ratio)
+        f = open("tmp-rating.txt", "r")
+        for i in range(3,0,-1):
+            l = f.readline()
+            solver = l.split()[0]
+            if solver in ratio_dic.keys():
+                score = ratio_dic.get(solver) + i
+            else:
+                score = i
+            ratio_dic[solver] = score
+        dic[ratio] = ratio_dic
+        return dic
+
+    dic = { 1:{},
+            2:{},
+            3:{},
+            4:{},
+            5:{},
+            6:{},
+            7:{},
+            8:{},
+        }
+    
     try:
+        os.system("rm -r benchmark-folder/*")
         while 1:
-            if random.random() < 0.5:    # Mas variables que clausulas
-                n_var = random.randrange(1,999)
-                ratio = random.randrange(1,8)
-                n_clauses = n_var//ratio
-            else:                       # Mas clausulas que variables
-                n_clauses = random.randrange(1,999)
-                ratio = random.randrange(1,8)
-                n_var = n_clauses//ratio
+            n_var = random.randrange(1,999)
+            ratio = random.randrange(1,8)
+            n_clauses = n_var//ratio
             rnd_cnf = "./rnd-cnf-gen-satisfiable.sh "+str(n_var)+" "+str(n_clauses)+" 3"
-            subprocess.call(rnd_cnf, shell=True)
+            if subprocess.call(rnd_cnf, shell=True) == 0:
+                subprocess.call("./rate-solvers.sh")
+                os.system("rm -r benchmark-folder/*")
+                dic = read_tmp_rating(ratio ,dic)
+            else:
+                print ("NO SE ENCONTRO SATISF")
     except KeyboardInterrupt:
         pass
-    
-    # Rate CNFs
-    cnfs = {1:[],2:[],3:[],4:[],5:[],6:[],7:[],8:[]} # Coje los resultados y con pandas los guarda en un csv
-    pdcnf = pd.DataFrame(cnfs, index=['1','2','3','4','5','6','7','8'])
-    
+    os.system("rm -f tmp-rating.txt")
+    os.system("rm -r benchmark-folder/*")
+    print("\n\n",dic)
+    with open("gotzilla-train.json","w") as j:
+        json.dump( dic, j)
 
 def parse(filename):
     clauses = []
